move streamverifier to its own file
